Added debugging information to get a sense of the magntitude of matrix values, and did a few bits of bugtesting along the way.

+++ src/kernels/embedding_layer.cu
@@ -2,6 +2,7 @@
 
 #include <kernels/matrix_device_kernels.cuh>
 
+
 // static void positional_encoding(matrix& input) {
 //     for (size_t token_i = 0; token_i < input.rows; ++token_i) {
 //         for (size_t encoding_i = 0; encoding_i < input.cols / 2;
@@ -23,10 +24,13 @@ static __global__ void positional_encoding_kernel(matrix_view input) {
     if (token_i < input.rows && encoding_i < input.cols / 2) {
         const auto offset = encoding_i * 2;
         const auto inner
-            = token_i / powf(10000.0f, offset / static_cast<float>(input.cols));
-        kernel::matrix::device_offset_elem(input, token_i, offset, sinf(inner));
+            = token_i
+              / std::pow(10000.0f, offset / static_cast<float>(input.cols));
+
+        kernel::matrix::device_offset_elem(input, token_i, offset,
+                                           std::sin(inner));
         kernel::matrix::device_offset_elem(input, token_i, offset + 1,
-                                           cosf(inner));
+                                           std::cos(inner));
     }
 }
